lots

+++ Assets/Scripts/Stats/Stat.cs
@@ -7,9 +7,9 @@ using UnityEngine;
 public class Stat
 {
     [SerializeField]
-    private int baseValue;
+    private float baseValue;
 
-    public int GetValue()
+    public float GetValue()
     {
         return baseValue;
     }
